+++ Dynamic/Types/DropdownType.php
@@ -44,7 +44,7 @@ class DropdownType implements FormFieldTypeInterface
     public function build(FormBuilderInterface $builder, FormField $field, $locale, $options)
     {
         $translation = $field->getTranslation($locale);
-        $options = array_merge($options, $this->getChoiceOptions($translation));
+        $options = array_merge($options, $this->getChoiceOptions($translation, $options['required']));
         $type = ChoiceType::class;
         $builder->add($field->getKey(), $type, $options);
     }